# src/main/resources/docs/openapi/REDAY_openapi_v1.yaml
openapi: 3.1.0
info:
  title: RE:DAY API
  version: 1.0.0
  description: 'RE:DAY 백엔드 API 문서


    응답 정책:

    - 일반 API는 성공/일반 실패 모두 HTTP 200을 사용하고, 본문의 success/code/message로 결과를 구분합니다.

    - JWT 인증 관련 실패만 HTTP 401/403을 사용합니다.

    - 각 엔드포인트의 200 응답에는 가능한 일반 실패 예시를 examples로 포함했습니다.'
servers:
- url: /api/v1
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        code:
          type: string
        message:
          type: string
        data: {}
      required:
      - success
      - code
      - message
      - data
...
  responses:
    UnauthorizedError:
      description: JWT 인증 실패
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          example:
            success: false
            code: COMMON_UNAUTHORIZED_FAIL
            message: 인증이 필요합니다.
            data: null
    ForbiddenError:
      description: 접근 권한 없음
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          example:
            success: false
            code: COMMON_FORBIDDEN_FAIL
            message: 접근 권한이 없습니다.
            data: null
...
paths:
...
  /schedules:
    get:
      summary: 일정 목록 조회
      security:
      - bearerAuth: []
      parameters:
      - name: viewType
        in: query
        required: true
        schema:
          type: string
          enum:
          - WEEK
          - MONTH
      - name: startDate
        in: query
        required: true
        schema:
          type: string
          description: yyyy-MM-dd
      - name: endDate
        in: query
        required: true
        schema:
          type: string
          description: yyyy-MM-dd
      responses:
        '200':
          description: 요청 성공 또는 일반 실패
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              examples:
                SCHEDULE_LIST_SUCCESS:
                  summary: 일정 목록 조회 성공
                  value:
                    success: true
                    code: SCHEDULE_LIST_SUCCESS
                    message: 일정 목록 조회에 성공했습니다.
                    data:
                      viewType: WEEK
                      startDate: '2026-01-05'
                      endDate: '2026-01-11'
                      schedules:
                      - scheduleId: 101
                        title: 운동하기
                        startAt: '2026-01-09 08:00:00'
                        estimatedMinutes: 15
                        actualMinutes: 20
                        status: DONE
                        completedAt: '2026-01-09 08:25:00'
                        deferCount: 0
                      - scheduleId: 102
                        title: NCS 문제 풀기
                        startAt: '2026-01-09 13:30:00'
                        estimatedMinutes: 60
                        actualMinutes: null
                        status: PENDING
                        completedAt: null
                        deferCount: 1
                SCH_INVALID_DATE_RANGE_FAIL:
                  summary: 조회 기간이 올바르지 않습니다.
                  description: startDate와 endDate가 없거나 범위가 잘못된 경우
                  value:
                    success: false
                    code: SCH_INVALID_DATE_RANGE_FAIL
                    message: 조회 기간이 올바르지 않습니다.
                    data: null
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
      description: 주간/월간 화면에 필요한 일정 목록을 조회합니다.
    post:
      summary: 일정 생성
      security:
      - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                startAt:
                  type: string
                  description: yyyy-MM-dd HH:mm:ss
                estimatedMinutes:
                  type: integer
                memo:
                  type:
                  - string
                  - 'null'
              required:
              - title
              - startAt
              - estimatedMinutes
      responses:
        '200':
          description: 요청 성공 또는 일반 실패
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              examples:
                SCHEDULE_CREATED:
                  summary: 일정 생성 성공
                  value:
                    success: true
                    code: SCHEDULE_CREATED
                    message: 일정이 등록되었습니다.
                    data:
                      scheduleId: 101
                SCH_CREATE_FAIL:
                  summary: 일정 생성에 실패하였습니다.
                  description: 일정 생성 처리 중 알 수 없는 오류가 발생한 경우
                  value:
                    success: false
                    code: SCH_CREATE_FAIL
                    message: 일정 생성에 실패하였습니다.
                    data: null
                SCH_INVALID_TITLE_FAIL:
                  summary: 일정 제목이 올바르지 않습니다.
                  description: 제목이 비어 있거나 허용 길이를 초과한 경우
                  value:
                    success: false
                    code: SCH_INVALID_TITLE_FAIL
                    message: 일정 제목이 올바르지 않습니다.
                    data: null
                SCH_INVALID_START_AT_FAIL:
                  summary: 시작 일시가 올바르지 않습니다.
                  description: 시작 일시 형식이 잘못되었거나 필수값이 누락된 경우
                  value:
                    success: false
                    code: SCH_INVALID_START_AT_FAIL
                    message: 시작 일시가 올바르지 않습니다.
                    data: null
                SCH_INVALID_ESTIMATED_MINUTES_FAIL:
                  summary: 예상 시간이 올바르지 않습니다.
                  description: 예상 시간이 0 이하이거나 허용 범위를 벗어난 경우
                  value:
                    success: false
                    code: SCH_INVALID_ESTIMATED_MINUTES_FAIL
                    message: 예상 시간이 올바르지 않습니다.
                    data: null
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
      description: 새 일정을 생성합니다.
  /schedules/{scheduleId}:
    get:
      summary: 일정 상세 조회
      security:
      - bearerAuth: []
      parameters:
      - name: scheduleId
        in: path
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: 요청 성공 또는 일반 실패
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              examples:
                SCHEDULE_DETAIL_SUCCESS:
                  summary: 일정 상세 조회 성공
                  value:
                    success: true
                    code: SCHEDULE_DETAIL_SUCCESS
                    message: 일정 상세 조회에 성공했습니다.
                    data:
                      scheduleId: 101
                      title: 운동하기
                      startAt: '2026-01-09 08:00:00'
                      estimatedMinutes: 15
                      actualMinutes: 20
                      memo: 아침 스트레칭
                      status: DONE
                      completedAt: '2026-01-09 08:25:00'
                      createdAt: '2026-01-08 22:10:00'
                      updatedAt: '2026-01-09 08:25:00'
                      deferCount: 1
                      deferLogs:
                      - actionLogId: 1001
                        actionType: DEFERRED
                        deferReasonCode: NO_TIME
                        deferReasonDetail: null
                        actionAt: '2026-01-09 07:55:00'
                SCH_NOT_FOUND_FAIL:
                  summary: 일정을 찾을 수 없습니다.
                  description: scheduleId에 해당하는 일정이 존재하지 않는 경우
                  value:
                    success: false
                    code: SCH_NOT_FOUND_FAIL
                    message: 일정을 찾을 수 없습니다.
                    data: null
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
      description: 일정 상세 정보와 미루기 로그를 조회합니다.
    patch:
      summary: 일정 수정
      security:
      - bearerAuth: []
      parameters:
      - name: scheduleId
        in: path
        required: true
        schema:
          type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                startAt:
                  type: string
                  description: yyyy-MM-dd HH:mm:ss
                estimatedMinutes:
                  type: integer
                memo:
                  type:
                  - string
                  - 'null'
      responses:
        '200':
          description: 요청 성공 또는 일반 실패
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              examples:
                SCHEDULE_UPDATED:
                  summary: 일정 수정 성공
                  value:
                    success: true
                    code: SCHEDULE_UPDATED
                    message: 일정이 수정되었습니다.
                    data: null
                SCH_UPDATE_FAIL:
                  summary: 일정 수정에 실패하였습니다.
                  description: 일정 수정 처리 중 알 수 없는 오류가 발생한 경우
                  value:
                    success: false
                    code: SCH_UPDATE_FAIL
                    message: 일정 수정에 실패하였습니다.
                    data: null
                SCH_NOT_FOUND_FAIL:
                  summary: 일정을 찾을 수 없습니다.
                  description: 수정 대상 일정이 존재하지 않는 경우
                  value:
                    success: false
                    code: SCH_NOT_FOUND_FAIL
                    message: 일정을 찾을 수 없습니다.
                    data: null
                SCH_INVALID_ESTIMATED_MINUTES_FAIL:
                  summary: 예상 시간이 올바르지 않습니다.
                  description: 예상 시간이 0 이하이거나 허용 범위를 벗어난 경우
                  value:
                    success: false
                    code: SCH_INVALID_ESTIMATED_MINUTES_FAIL
                    message: 예상 시간이 올바르지 않습니다.
                    data: null
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
      description: 제목, 시작일시, 예상시간, 메모를 수정합니다.
    delete:
      summary: 일정 삭제
      security:
      - bearerAuth: []
      parameters:
      - name: scheduleId
        in: path
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: 요청 성공 또는 일반 실패
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              examples:
                SCHEDULE_DELETED:
                  summary: 일정 삭제 성공
                  value:
                    success: true
                    code: SCHEDULE_DELETED
                    message: 일정이 삭제되었습니다.
                    data: null
                SCH_DELETE_FAIL:
                  summary: 일정 삭제에 실패하였습니다.
                  description: 일정 삭제 처리 중 알 수 없는 오류가 발생한 경우
                  value:
                    success: false
                    code: SCH_DELETE_FAIL
                    message: 일정 삭제에 실패하였습니다.
                    data: null
                SCH_NOT_FOUND_FAIL:
                  summary: 일정을 찾을 수 없습니다.
                  description: 삭제 대상 일정이 존재하지 않는 경우
                  value:
                    success: false
                    code: SCH_NOT_FOUND_FAIL
                    message: 일정을 찾을 수 없습니다.
                    data: null
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
      description: 일정을 삭제합니다. deletedAt을 설정하는 soft delete 방식입니다.
  /schedules/{scheduleId}/complete:
    post:
      summary: 일정 완료 처리
      security:
      - bearerAuth: []
      parameters:
      - name: scheduleId
        in: path
        required: true
        schema:
          type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                actualMinutes:
                  type: integer
              required:
              - actualMinutes
      responses:
        '200':
          description: 요청 성공 또는 일반 실패
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              examples:
                SCHEDULE_COMPLETED:
                  summary: 일정 완료 성공
                  value:
                    success: true
                    code: SCHEDULE_COMPLETED
                    message: 일정이 완료 처리되었습니다.
                    data:
                      scheduleId: 101
                      status: DONE
                      actualMinutes: 25
                      completedAt: '2026-01-09 08:25:00'
                SCH_COMPLETE_FAIL:
                  summary: 일정 완료 처리에 실패하였습니다.
                  description: 일정 완료 처리 중 알 수 없는 오류가 발생한 경우
                  value:
                    success: false
                    code: SCH_COMPLETE_FAIL
                    message: 일정 완료 처리에 실패하였습니다.
                    data: null
                SCH_NOT_FOUND_FAIL:
                  summary: 일정을 찾을 수 없습니다.
                  description: 완료 처리 대상 일정이 존재하지 않는 경우
                  value:
                    success: false
                    code: SCH_NOT_FOUND_FAIL
                    message: 일정을 찾을 수 없습니다.
                    data: null
                SCH_ALREADY_DONE_FAIL:
                  summary: 이미 완료된 일정입니다.
                  description: 상태가 DONE인 일정을 다시 완료 처리하려는 경우
                  value:
                    success: false
                    code: SCH_ALREADY_DONE_FAIL
                    message: 이미 완료된 일정입니다.
                    data: null
                SCH_INVALID_ACTUAL_MINUTES_FAIL:
                  summary: 실제 소요 시간이 올바르지 않습니다.
                  description: 실제 소요 시간이 0 이하이거나 허용 범위를 벗어난 경우
                  value:
                    success: false
                    code: SCH_INVALID_ACTUAL_MINUTES_FAIL
                    message: 실제 소요 시간이 올바르지 않습니다.
                    data: null
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
      description: 일정을 완료 처리하고 실제 소요 시간을 저장합니다.
  /schedules/{scheduleId}/defer:
    post:
      summary: 일정 미루기 처리
      security:
      - bearerAuth: []
      parameters:
      - name: scheduleId
        in: path
        required: true
        schema:
          type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                deferReasonCode:
                  type: string
                deferReasonDetail:
                  type:
                  - string
                  - 'null'
                newStartAt:
                  type:
                  - string
                  - 'null'
                  description: yyyy-MM-dd HH:mm:ss
              required:
              - deferReasonCode
      responses:
        '200':
          description: 요청 성공 또는 일반 실패
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              examples:
                SCHEDULE_DEFERRED:
                  summary: 일정 미루기 성공
                  value:
                    success: true
                    code: SCHEDULE_DEFERRED
                    message: 일정이 미뤄졌습니다.
                    data:
                      scheduleId: 102
                      status: PENDING
                      startAt: '2026-01-09 20:00:00'
                      deferCount: 2
                SCH_DEFER_FAIL:
                  summary: 일정 미루기에 실패하였습니다.
                  description: 일정 미루기 처리 중 알 수 없는 오류가 발생한 경우
                  value:
                    success: false
                    code: SCH_DEFER_FAIL
                    message: 일정 미루기에 실패하였습니다.
                    data: null
                SCH_NOT_FOUND_FAIL:
                  summary: 일정을 찾을 수 없습니다.
                  description: 미루기 대상 일정이 존재하지 않는 경우
                  value:
                    success: false
                    code: SCH_NOT_FOUND_FAIL
                    message: 일정을 찾을 수 없습니다.
                    data: null
                SCH_ALREADY_DONE_FAIL:
                  summary: 이미 완료된 일정입니다.
                  description: 완료된 일정에 대해 미루기 요청한 경우
                  value:
                    success: false
                    code: SCH_ALREADY_DONE_FAIL
                    message: 이미 완료된 일정입니다.
                    data: null
                SCH_INVALID_DEFER_REASON_FAIL:
                  summary: 미루기 사유가 올바르지 않습니다.
                  description: 허용되지 않은 미루기 사유 코드가 전달된 경우
                  value:
                    success: false
                    code: SCH_INVALID_DEFER_REASON_FAIL
                    message: 미루기 사유가 올바르지 않습니다.
                    data: null
                SCH_DEFER_REASON_DETAIL_REQUIRED_FAIL:
                  summary: 미루기 상세 사유를 입력해주세요.
                  description: CUSTOM 사유인데 상세 사유가 비어 있는 경우
                  value:
                    success: false
                    code: SCH_DEFER_REASON_DETAIL_REQUIRED_FAIL
                    message: 미루기 상세 사유를 입력해주세요.
                    data: null
                SCH_INVALID_NEW_START_AT_FAIL:
                  summary: 변경할 시작 일시가 올바르지 않습니다.
                  description: newStartAt 형식이 잘못되었거나 유효하지 않은 경우
                  value:
                    success: false
                    code: SCH_INVALID_NEW_START_AT_FAIL
                    message: 변경할 시작 일시가 올바르지 않습니다.
                    data: null
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
      description: 미루기 사유를 저장하고 필요 시 시작일시를 변경합니다.
...